// interfaceses/country.interface.ts
import ICapitalInfo from "./capitalInfo.interface";
import ICar from "./car.interface";
import ICoatOfArms from "./coatOfArms.interface";
import ICountryName from "./countryName.interface";
import ICurrency from "./currency.interface";
import IDemonyms from "./demonyms.interface";
import IFlags from "./flags.interface";
import IIDD from "./IDD.interface";
import IMaps from "./maps.interface";
import IPostalCode from "./postalCode.interface";
import ITranslation from "./translation.interface";

export default interface ICountry {
  name: ICountryName;
  tld: string[];
  cca2: string;
  ccn3: string;
  cca3: string;
  cioc: string;
  independent: boolean;
  status: string;
  unMember: boolean;
  currencies: Record<string, ICurrency>;
  idd: IIDD;
  capital: string[];
  altSpellings: string[];
  region: string;
  subregion: string;
  languages: Record<string, string>;
  translations: Record<string, ITranslation>;
  latlng: number[];
  landlocked: boolean;
  borders: string[];
  area: number;
  demonyms: IDemonyms;
  flag: string;
  maps: IMaps;
  population: number;
  gini: Record<string, number>;
  fifa: string;
  car: ICar;
  timezones: string[];
  continents: string[];
  flags: IFlags;
  coatOfArms: ICoatOfArms;
  startOfWeek: string;
  capitalInfo: ICapitalInfo;
  postalCode: IPostalCode;
}